+++ CHANGELOG.md
@@ -52,6 +52,7 @@ v1.8.0 [unreleased]
 - [#333](https://github.com/latera/homs/pull/333) Optimize Camunda process definitions fetching.
 - [#332](https://github.com/latera/homs/pull/332) Add table for claiming.
 - [#337](https://github.com/latera/homs/pull/337) Add claim button for widget page.
+- [#338](https://github.com/latera/homs/pull/338) Add task overview component.
 
 v1.7.8 [2019-10-03]
 -------------------